move notebook controller and serializer to common location
