+++ src/polyglot-notebooks-vscode/.vscode/launch.json
@@ -6,17 +6,6 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "type": "pwa-node",
-      "request": "launch",
-      "name": "Debug Mocha all tests",
-      "program": "${workspaceFolder}/node_modules/mocha/bin/_mocha",
-      "args": [
-        "${workspaceRoot}/tests/**/*.spec.ts"
-      ],
-      "cwd": "${workspaceRoot}",
-      "internalConsoleOptions": "openOnSessionStart"
-    },
-    {
       "name": "Run Extension",
       "type": "extensionHost",
       "request": "launch",
